# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API documentation for the Image Processing Service project.",
        "title": "Image Processing Service API",
        "contact": {
            "name": "Image Processing Service Team",
            "email": "dev@image-processing.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/health": {
            "get": {
                "description": "Returns the current API health status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Health check",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Trigger validation error example",
                        "name": "fail",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HealthSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/v1/images/upload": {
            "post": {
                "description": "Accepts a multipart image file, validates it, uploads it to MinIO, and persists metadata to PostgreSQL.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Images"
                ],
                "summary": "Upload image and persist metadata",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UploadSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.UploadResponse": {
            "type": "object",
            "properties": {
                "content_type": {
                    "type": "string",
                    "example": "image/jpeg"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "object_key": {
                    "type": "string",
                    "example": "images/550e8400-e29b-41d4-a716-446655440000.jpg"
                },
                "size": {
                    "type": "integer",
                    "example": 12345
                },
                "status": {
                    "type": "string",
                    "example": "uploaded"
                }
            }
        },
        "dto.UploadSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.UploadResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Image uploaded successfully"
                },
                "meta": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.ErrorItem"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Validation failed"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "response.ErrorItem": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "field": {
                    "type": "string",
                    "example": "file"
                },
                "message": {
                    "type": "string",
                    "example": "file is required"
                }
            }
        },
        "response.HealthData": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "healthy"
                }
            }
        },
        "response.HealthSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/response.HealthData"
                },
                "message": {
                    "type": "string",
                    "example": "Image Processing Service API is running"
                },
                "meta": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    }
}